Changed int to long

+++ Main.java
@@ -1,10 +1,10 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {300, 400, 500, 200};
+        long[] sales = {300, 400, 500, 200};
 
         SalesManager salesManager = new SalesManager(sales);
 
-        int result = salesManager.max();
+        long result = salesManager.max();
         System.out.println(result);
     }
 }
